* develop: (22 commits)
  added execution to build the documentation
  updated docs for newer rst-theme versions
  added query parameter check
  throw proper exception on wrong formatted filters
  fixed missing fields for test
  added test for fetch groups
  added missing jackson-databind dependency
  improved field parsing to properly handle "from" and "to" fields on relationships
  changed QueryBuilder invocation to always create a new one
  added method to get a stream of results
  fixed sum method wrong cast
  improved QueryField parsing to include sub-fields
  introducing fetch groups
  some minor code improvements
  improved comments
  improved exception message
  prevent index dropping when there's a constraint
  upgraded kotlin to version 1.2.41
  fixed duplicate field entries in model cache fixed project version in pom's
  updated kotlin version
  ...

# Conflicts:
#	auth/pom.xml
#	base/pom.xml
#	core/pom.xml
#	examples/basic/pom.xml
#	examples/custom-service/pom.xml
#	examples/oauth/pom.xml
#	pom.xml
#	test/pom.xml

+++ base/src/main/java/de/whitefrog/frogr/Service.java
@@ -249,6 +249,8 @@ public class Service implements AutoCloseable {
         if(!Base.class.isAssignableFrom(modelClass)) 
           throw new FrogrException("model class " + modelClass.getName() + " is not of type Base");
 
+        if(logger.isDebugEnabled()) logger.debug("creating constraints for {}", modelClass.getSimpleName());
+
         ModelRepository repository = (ModelRepository) repository(modelClass);
         List<ConstraintDefinition> constraints = Iterables.asList(
           schema.getConstraints(repository.label()));
@@ -295,7 +297,7 @@ public class Service implements AutoCloseable {
               .create();
             logger.debug("created {} index on field \"{}\" for model \"{}\"",
               annotations.indexed.type(), descriptor.getName(), repository.getModelClass().getSimpleName());
-          } else if(annotations.indexed == null && !annotations.unique && existingIndex != null) {
+          } else if(annotations.indexed == null && !annotations.unique && existingIndex != null && existingConstraint == null) {
             existingIndex.drop();
             logger.debug("dropped index on field \"{}\" for model \"{}\"",
               descriptor.getName(), repository.getModelClass().getSimpleName());
@@ -321,7 +323,7 @@ public class Service implements AutoCloseable {
 
   public void shutdown() {
     state = State.ShuttingDown;
-    repositoryFactory().cache().forEach(Repository::dispose);
+    if(repositoryFactory() != null) repositoryFactory().cache().forEach(Repository::dispose);
     if(graphDb != null) graphDb.shutdown();
     state = State.Started;
   }
